refactor: extract addImportsToNgModule rule
